Настройка CI

+++ test/functional_tests/functional.cpp
@@ -54,7 +54,7 @@ TEST_F(Functional_test, HandlesFirstRegularCase) {
     EXPECT_TRUE(output_static_stream != NULL);
     EXPECT_TRUE(output_dynamic_stream != NULL);
 
-    int capacity_of_employees = dyn_lib.read_number(input_stream);
+    int capacity_of_employees = (int)dyn_lib.read_number(input_stream);
     /*if (capacity_of_employees != FAILURE) {
         employee_info **employees = dyn_lib->read_employees(input_stream, capacity_of_employees);
